pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "AliJaddi"
version = "0.5.1-beta"
description = "AliJaddi — منصة نماذج الذكاء الاصطناعي (Qt for Python)"
requires-python = ">=3.10"
dependencies = [
    "PySide6-Essentials>=6.6.0",
    "httpx>=0.28.0",
    "python-dotenv>=1.0.0",
    "requests>=2.28.0",
]

[project.optional-dependencies]
build = ["pyinstaller>=6.0.0"]
android = ["PySide6>=6.6.0"]
dev = ["pytest>=7.0.0", "fastapi>=0.109.0"]
api = ["fastapi>=0.109.0", "uvicorn[standard]>=0.27.0"]
bentoml = ["bentoml>=1.2.0"]

[project.scripts]
alijaddi = "main_qt:main"

# منع اكتشاف المجلدات العلوية تلقائياً (ui، services، تنزيل، …) — يفشل التثبيت القابل للتحرير بدون هذا.
[tool.setuptools]
packages = ["alijaddi"]
py-modules = ["main_qt", "launch", "Ali12"]

[tool.pyinstaller]
entry = "main_qt.py"
name = "AliJaddi"
icon = "icon.png"
windowed = true
